Added single field reference list validator to SingleFieldValidator.
