+++ app.py
@@ -2,9 +2,9 @@ import os
 
 from flask import Flask
 
-from mlrvalidator.validators.error_validator import ErrorValidator
-from mlrvalidator.validators.single_field_warning_validator import SingleFieldWarningValidator
-from mlrvalidator.schema import warning_schema
+#from mlrvalidator.validators.error_validator import ErrorValidator
+#from mlrvalidator.validators.single_field_warning_validator import SingleFieldWarningValidator
+#from mlrvalidator.schema import warning_schema
 
 application = Flask(__name__)
 
@@ -14,11 +14,11 @@ PROJECT_DIR = os.path.dirname(__file__)
 if os.path.exists(os.path.join(PROJECT_DIR, '.env')):
     application.config.from_pyfile('.env')
 
-error_validator = ErrorValidator()
-warning_validator = SingleFieldWarningValidator(warning_schema, allow_unknown=True)
+#error_validator = ErrorValidator()
+#warning_validator = SingleFieldWarningValidator(warning_schema, allow_unknown=True)
 
 
-from mlrvalidator.services import *
+#from mlrvalidator.services import *
 
 if __name__ == '__main__':
     application.run()
