+++ mlrvalidator/schemas/single_field_validation_schema.yml
@@ -2,6 +2,7 @@ agencyCode:
     required: True
     is_empty: False
     maxlength: 5
+    valid_reference: True
 siteNumber:
     required: True
     maxlength: 15
@@ -10,11 +11,12 @@ stationName:
     required: True
     maxlength: 50
     is_empty: False
-    valid_special_chars: True
+    regex: ^[^\t#\*\\"\^_\$]*$
 siteTypeCode:
     required: True
     is_empty: False
     maxlength: 7
+    valid_reference: True
 latitude:
     maxlength: 11
     valid_latitude_dms: True
@@ -23,18 +25,23 @@ longitude:
     valid_longitude_dms: True
 coordinateAccuracyCode:
     maxlength: 1
+    valid_reference: True
 coordinateDatumCode:
     maxlength: 10
+    valid_reference: True
 coordinateMethodCode:
     maxlength: 1
+    valid_reference: True
 altitude:
     type: numeric
     maxlength: 8
     valid_precision: True
 altitudeDatumCode:
     maxlength: 10
+    valid_reference: True
 altitudeMethodCode:
     maxlength: 1
+    valid_reference: True
 altitudeAccuracyValue:
     type: numeric
     maxlength: 3
@@ -42,10 +49,12 @@ districtCode:
     required: True
     is_empty: False
     maxlength: 3
+    valid_reference: True
 countryCode:
     required: True
     is_empty: False
     maxlength: 2
+    valid_reference: True
 stateFipsCode:
     required: True
     is_empty: False
@@ -66,10 +75,13 @@ aquiferCode:
     maxlength: 8
 aquiferTypeCode:
     maxlength: 1
+    valid_reference: True
 agencyUseCode:
     maxlength: 1
+    valid_reference: True
 dataReliabilityCode:
     maxlength: 1
+    valid_reference: True
 landNet:
     maxlength: 23
     valid_land_net: True
@@ -82,24 +94,31 @@ nationalWaterUseCode:
     maxlength: 2
 primaryUseOfSite:
     maxlength: 1
+    valid_reference: True
 secondaryUseOfSite:
     maxlength: 1
+    valid_reference: True
 tertiaryUseOfSiteCode:
     maxlength: 1
+    valid_reference: True
 primaryUseOfWaterCode:
     maxlength: 1
+    valid_reference: True
 secondaryUseOfWaterCode:
     maxlength: 1
+    valid_reference: True
 tertiaryUseOfWaterCode:
     maxlength: 1
+    valid_reference: True
 topographicCode:
     maxlength: 1
+    valid_reference: True
 dataTypesCode:
     maxlength: 30
-    valid_data_types_chars: True
+    regex: ^[^AION ]*$
 instrumentsCode:
     maxlength: 30
-    valid_instruments_chars: True
+    regex: ^[^YN ]*$
 contributingDrainageArea:
     type: positive_numeric
     maxlength: 8
@@ -127,6 +146,7 @@ timeZoneCode:
     required: True
     is_empty: False
     maxlength: 5
+    valid_reference: True
 daylightSavingsTimeFlag:
     required: True
     is_empty: False
@@ -136,5 +156,6 @@ remarks:
     maxlength: 50
 siteWebReadyCode:
     maxlength: 1
+    valid_reference: True
 gwFileCode:
     maxlength: 20
