+++ mlrvalidator/validators/tests/test_cross_field_validator.py
@@ -212,3 +212,57 @@ class CrossFieldValidatorCheckValidSiteTypeUpdate(TestCase):
 
     def test_invalid_site_type_on_update(self):
         self.assertFalse(self.validator.validate({'siteTypeCode': 'FA-SPS'}, {'siteTypeCode': 'GW-MW'}, update=True))
+
+
+class ValidateValidLandNetTestCase(TestCase):
+
+    def setUp(self):
+        self.validator = SingleFieldValidator(schema={'landNet': {'valid_land_net': True}})
+        self.good_data = {
+            'landNet': 'SWSWSWS010T09832R093425'
+        }
+        self.good_data2 = {
+            'landNet': '      S15 T20N  R11E'
+        }
+        self.good_data3 = {
+            'landNet': 'NWNWSWS15 T014N R022E 4'
+        }
+        self.good_data4 = {
+            'landNet': '      S   T23N  R20E  4'
+        }
+        self.bad_data = {
+            'landNet': 'Q'
+        }
+        self.bad_data2 = {
+            'landNet': 'NWNWSWS15 T014N R02-E 4'
+        }
+        self.bad_data3 = {
+            'landNet': 'NWNWSWS15  T014N R022E4'
+        }
+        self.bad_data4 = {
+            'landNet': 'NWNWSW S15 T014N R022E4'
+        }
+        self.bad_data5 = {
+            'landNet': 'NWNWSWF15 T014N R 022E4'
+        }
+        self.bad_data6 = {
+            'landNet': 'NWNWSWS15 S014N R 022E4'
+        }
+        self.bad_data7 = {
+            'landNet': 'NWNWSWF15 T014N S 022E4'
+        }
+
+    def test_validate_ok(self):
+        self.assertTrue(self.validator.validate(self.good_data))
+        self.assertTrue(self.validator.validate(self.good_data2))
+        self.assertTrue(self.validator.validate(self.good_data3))
+        self.assertTrue(self.validator.validate(self.good_data4))
+
+    def test_with_validate_not_ok(self):
+        self.assertFalse(self.validator.validate(self.bad_data))
+        self.assertFalse(self.validator.validate(self.bad_data2))
+        self.assertFalse(self.validator.validate(self.bad_data3))
+        self.assertFalse(self.validator.validate(self.bad_data4))
+        self.assertFalse(self.validator.validate(self.bad_data5))
+        self.assertFalse(self.validator.validate(self.bad_data6))
+        self.assertFalse(self.validator.validate(self.bad_data7))
