+++ mlrvalidator/validators/tests/test_reference_validator.py
@@ -7,219 +7,6 @@ site_validator = ReferenceValidator(reference_schema)
 site_validator.allow_unknown = True
 
 
-class ValidateReferenceCase(TestCase):
-    #TODO: Turn these into true unit tests by specifying the schema in the test case and mocking the reference lists
-    #TODO: Add some tests that use an existing document.
-
-    def setUp(self):
-        self.good_data = {
-            'agencyCode': 'USGS'
-        }
-        self.good_data2 = {
-            'agencyCode': 'USGS '
-        }
-        self.good_data3 = {
-            'agencyUseCode': 'A'
-        }
-        self.good_data4 = {
-            'altitudeDatumCode': 'LMSL'
-        }
-        self.good_data5 = {
-            'altitudeMethodCode': 'A'
-        }
-        self.good_data6 = {
-            'altitudeMethodCode': ' '
-        }
-        self.good_data7 = {
-            'altitudeMethodCode': ''
-        }
-        self.good_data8 = {
-            'aquiferTypeCode': 'C'
-        }
-        self.good_data9 = {
-            'coordinateDatumCode': 'ACCRA'
-        }
-        self.good_data10 = {
-            'coordinateMethodCode': 'C'
-        }
-        self.good_data11 = {
-            'coordinateAccuracyCode': '1'
-        }
-        self.good_data12 = {
-            'coordinateAccuracyCode': ' '
-        }
-        self.good_data13 = {
-            'coordinateAccuracyCode': ''
-        }
-        self.good_data14 = {
-            'countryCode': 'US'
-        }
-        self.good_data15 = {
-            'dataReliabilityCode': 'C'
-        }
-        self.good_data16 = {
-            'districtCode': '01'
-        }
-        self.good_data17 = {
-            'primaryUseOfSite': 'A'
-        }
-        self.good_data18 = {
-            'primaryUseOfWaterCode': 'A'
-        }
-        self.good_data19 = {
-            'secondaryUseOfSite': 'A'
-        }
-        self.good_data20 = {
-            'secondaryUseOfWaterCode': 'A'
-        }
-        self.good_data21 = {
-            'siteTypeCode': 'AG'
-        }
-        self.good_data22 = {
-            'siteWebReadyCode': 'Y'
-        }
-        self.good_data23 = {
-            'tertiaryUseOfSiteCode': 'A'
-        }
-        self.good_data24 = {
-            'tertiaryUseOfWaterCode': 'A'
-        }
-        self.good_data25 = {
-             'timeZoneCode': 'CST'
-        }
-        self.good_data26 = {
-            'topographicCode': 'A'
-        }
-        self.good_data27 = {
-            'topographicCode': 'a'
-        }
-        self.good_data28 = {
-            'sourceOfDepthCode': 'a'
-        }
-        self.good_data29 = {
-            'sourceOfDepthCode': 'A'
-        }
-        self.bad_data = {
-            'agencyCode': 'x'
-        }
-        self.bad_data2 = {
-            'agencyUseCode': 'XYZ'
-        }
-        self.bad_data3 = {
-            'altitudeDatumCode': 'XYZ'
-        }
-        self.bad_data4 = {
-            'altitudeMethodCode': 'B'
-        }
-        self.bad_data5 = {
-            'aquiferTypeCode': 'B'
-        }
-        self.bad_data6 = {
-            'coordinateDatumCode': 'XYZ'
-        }
-        self.bad_data7 = {
-            'coordinateMethodCode': 'A'
-        }
-        self.bad_data8 = {
-            'coordinateAccuracyCode': 'A'
-        }
-        self.bad_data9 = {
-            'countryCode': 'XY'
-        }
-        self.bad_data10 = {
-            'dataReliabilityCode': 'A'
-        }
-        self.bad_data11 = {
-            'districtCode': '99'
-        }
-        self.bad_data12 = {
-            'primaryUseOfSite': 'B'
-        }
-        self.bad_data13 = {
-            'primaryUseOfWaterCode': 'V'
-        }
-        self.bad_data14 = {
-            'secondaryUseOfSite': 'B'
-        }
-        self.bad_data15 = {
-            'secondaryUseOfWaterCode': 'V'
-        }
-        self.bad_data16 = {
-            'siteTypeCode': 'AA'
-        }
-        self.bad_data17 = {
-            'siteWebReadyCode': 'F'
-        }
-        self.bad_data18 = {
-            'tertiaryUseOfSiteCode': 'B'
-        }
-        self.bad_data19 = {
-            'tertiaryUseOfWaterCode': 'V'
-        }
-        self.bad_data20 = {
-            'timeZoneCode': 'XYZ'
-        }
-        self.bad_data21 = {
-            'topographicCode': 'I'
-        }
-        self.bad_data22 = {
-            'sourceOfDepthCode': 'X'
-        }
-
-    def test_validate_ok(self):
-        self.assertTrue(site_validator.validate(self.good_data, {}))
-        self.assertTrue(site_validator.validate(self.good_data2, {}))
-        self.assertTrue(site_validator.validate(self.good_data3, {}))
-        self.assertTrue(site_validator.validate(self.good_data4, {}))
-        self.assertTrue(site_validator.validate(self.good_data5, {}))
-        self.assertTrue(site_validator.validate(self.good_data6, {}))
-        self.assertTrue(site_validator.validate(self.good_data7, {}))
-        self.assertTrue(site_validator.validate(self.good_data8, {}))
-        self.assertTrue(site_validator.validate(self.good_data9, {}))
-        self.assertTrue(site_validator.validate(self.good_data10, {}))
-        self.assertTrue(site_validator.validate(self.good_data11, {}))
-        self.assertTrue(site_validator.validate(self.good_data12, {}))
-        self.assertTrue(site_validator.validate(self.good_data13, {}))
-        self.assertTrue(site_validator.validate(self.good_data14, {}))
-        self.assertTrue(site_validator.validate(self.good_data15, {}))
-        self.assertTrue(site_validator.validate(self.good_data16, {}))
-        self.assertTrue(site_validator.validate(self.good_data17, {}))
-        self.assertTrue(site_validator.validate(self.good_data18, {}))
-        self.assertTrue(site_validator.validate(self.good_data19, {}))
-        self.assertTrue(site_validator.validate(self.good_data20, {}))
-        self.assertTrue(site_validator.validate(self.good_data21, {}))
-        self.assertTrue(site_validator.validate(self.good_data22, {}))
-        self.assertTrue(site_validator.validate(self.good_data23, {}))
-        self.assertTrue(site_validator.validate(self.good_data24, {}))
-        self.assertTrue(site_validator.validate(self.good_data25, {}))
-        self.assertTrue(site_validator.validate(self.good_data26, {}))
-        self.assertTrue(site_validator.validate(self.good_data27, {}))
-        self.assertTrue(site_validator.validate(self.good_data28, {}))
-        self.assertTrue(site_validator.validate(self.good_data29, {}))
-
-    def test_with_validate_not_ok(self):
-        self.assertFalse(site_validator.validate(self.bad_data, {}))
-        self.assertFalse(site_validator.validate(self.bad_data2, {}))
-        self.assertFalse(site_validator.validate(self.bad_data3, {}))
-        self.assertFalse(site_validator.validate(self.bad_data4, {}))
-        self.assertFalse(site_validator.validate(self.bad_data5, {}))
-        self.assertFalse(site_validator.validate(self.bad_data6, {}))
-        self.assertFalse(site_validator.validate(self.bad_data7, {}))
-        self.assertFalse(site_validator.validate(self.bad_data8, {}))
-        self.assertFalse(site_validator.validate(self.bad_data9, {}))
-        self.assertFalse(site_validator.validate(self.bad_data10, {}))
-        self.assertFalse(site_validator.validate(self.bad_data11, {}))
-        self.assertFalse(site_validator.validate(self.bad_data12, {}))
-        self.assertFalse(site_validator.validate(self.bad_data13, {}))
-        self.assertFalse(site_validator.validate(self.bad_data14, {}))
-        self.assertFalse(site_validator.validate(self.bad_data15, {}))
-        self.assertFalse(site_validator.validate(self.bad_data16, {}))
-        self.assertFalse(site_validator.validate(self.bad_data17, {}))
-        self.assertFalse(site_validator.validate(self.bad_data18, {}))
-        self.assertFalse(site_validator.validate(self.bad_data19, {}))
-        self.assertFalse(site_validator.validate(self.bad_data20, {}))
-        self.assertFalse(site_validator.validate(self.bad_data21, {}))
-        self.assertFalse(site_validator.validate(self.bad_data22, {}))
 
 
 class ValidateAquiferCode(TestCase):
